feat: nickname 변경 api 구현 및 api 폴더 분리 [TRL-75]
