+++ src/api/index.ts
@@ -4,8 +4,9 @@ import {
   logout,
   refreshAccessToken,
   checkRefreshToken,
-  getUserProfile,
 } from '@/api/oauth';
+
+import { getUserProfile, changeNickname } from '@/api/userInfo';
 
 const HTTP = {
   getLoginUri,
@@ -14,6 +15,7 @@ const HTTP = {
   refreshAccessToken,
   checkRefreshToken,
   getUserProfile,
+  changeNickname,
 };
 
 export default HTTP;
