+++ src/api/oauth.ts
@@ -17,11 +17,6 @@ interface LogoutResponse {
   response?: boolean;
 }
 
-interface UserProfile {
-  nickname: string;
-  imgUrl: string;
-}
-
 // 백엔드 서버에 OAuth 로그인 url 요청
 export const getLoginUri = async (oauthServer: string) => {
   const res = await axios({
@@ -68,14 +63,5 @@ export const logout = async () => {
     requireAuth: false,
   });
   delete axios.defaults.headers.common.Authorization;
-  return res.data;
-};
-
-export const getUserProfile = async () => {
-  const res = await axios<UserProfile>({
-    method: 'get',
-    url: `/user-profile`,
-    requireAuth: true,
-  });
   return res.data;
 };
